Clean up code.
- Remove the implementation of the WideResNet from RobustBench.
- Move some functions in `rotta.py` to either `utils.py` or `RoTTA` class.`
